clarify what field is being made private
Great post!  Thanks, it helped me out.  This is clarifying one part I got caught up on, which was the paragraph about making "the field" private.  It took me a while to realize which field was being referred to.  It's clear now, but for a while I thought you meant the Angle struct itself should be made private.  Here's a clarification that would have helped me.  Thanks again!